Porting ac3d plugin to Linux

+++ src/DSF/DSFPointPool.cpp
@@ -552,7 +552,7 @@ void DSFOptimizePrimitives(
 			int offset = 0;
 			while (1)
 			{
-				int num = min(new_prim->m_Indices.size() - offset, 255UL);
+				int num = min(new_prim->m_Indices.size() - offset, (size_t)255);
 				if(num == 0) 
 					break;
 				out_prims.push_back(DSFPrimitive());
